HomeworkSeven

+++ src/homeworkFive/Mouse.java
@@ -2,7 +2,7 @@ package homeworkFive;
 
 public class Mouse extends Product {
 
-    public Mouse(int id, String name, double price){
+    public Mouse(int id, String name, int price){
         super(id, name, price);
     }
 }
